backend/services/tenant/pyproject.toml
[project]
name = "verdantiq-tenant"
version = "0.1.0"
description = "VerdantIQ Tenant & Billing Service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.7.1",
    "PyJWT>=2.10.0",
    "cryptography>=44.0.0",
    "python-dateutil>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py313"
